extract commands
So we can reuse them in other places
