+++ lib/baes/helpers.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# module to encapsulate helpers
+module Baes::Helpers
+end
+
+require_relative "helpers/commands"
